requirements.txt
astropy
matplotlib
numpy
Pillow
python-decouple
requests
git+https://gitlab.com/kerel-fs/python-satnogs-api.git@0.2
h5py~=3.6.0
pandas~=1.3.4
ephem~=4.1
scipy~=1.7.2
beyond~=0.7.3
mplcursors~=0.5.1
pyarrow
tables
